requirements.txt
altair==4.1.0
arabic-reshaper==3.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==4.1.0
blinker==1.5
cachetools==4.2.4
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
comm==0.1.4
commonmark==0.9.1
contrib==0.3.0
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
Django==3.2.25
django-crispy-forms==1.13.0
django-easy-pdf==0.1.1
django-jazzmin==2.6.1
django-widget-tweaks==1.4.11
djangorestframework==3.15.1
entrypoints==0.4
filelock==3.4.1
gitdb==4.0.9
GitPython==3.1.18
html5lib==1.1
idna==3.6
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.8.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.1
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.7
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mysql==0.0.3
mysqlclient==1.4.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.8
notebook==6.4.10
numpy==1.19.5
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prometheus-client==0.17.1
prompt-toolkit==3.0.36
protobuf==3.19.6
pyarrow==6.0.1
pycparser==2.21
pydeck==0.6.2
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.1.1
PyPDF2==3.0.1
pyrsistent==0.18.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==305
pywinpty==1.1.6
PyYAML==6.0.1
pyzmq==25.1.1
reportlab==3.6.8
requests==2.27.1
rich==12.6.0
scikit-learn==0.24.2
scipy==1.5.4
semver==2.13.0
Send2Trash==1.8.2
six==1.16.0
sklearn==0.0
smmap==5.0.0
sqlparse==0.3.1
streamlit==1.10.0
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.1
traitlets==4.3.3
typing_extensions==4.1.1
tzdata==2024.1
tzlocal==4.2
urllib3==1.26.18
validators==0.20.0
virtualenv==20.17.1
watchdog==2.3.1
wcwidth==0.2.9
webencodings==0.5.1
widgetsnbextension==3.6.6
WTForms==3.0.0
xhtml2pdf==0.2.5
zipp==3.6.0
